requirements.txt
mcp>=0.10.0
pydantic>=2.0.0
fpdf2>=2.7.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.0.0
python-pptx>=0.6.21
Pillow>=10.0.0
icalendar>=5.0.0
vobject>=0.9.6.1
pytest